requirements.txt
asgiref==3.7.2
bleach==6.1.0
Django==4.2.7
django-cors-headers==3.7.0
django-filter==23.3
django-post-office==3.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-writable-nested==0.7.0
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.2
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
webencodings==0.5.1
phonenumbers~=8.13.26